[kanren] List Mapping and Variables in Relations
Adds a function that allows mapping two lists to eachother via a goal applied to each element.
Also adds the ability for ecs::Relations to store list of terms (and thus logic variables) instead of just entities.
Also also adds a version of related_entities that deals with a list rather than a stream.